fix(PhysicsEngine): Repair 二进制“*”:“glm::vec<2,float,glm::packed_highp>”不定义该运算符或到预定义运算符可接收的类型的转换
速度、力和质量都是 float 类型，但是更新方法中的时间却是 double 类型

+++ src/engine/physics/physics_engine.cpp
@@ -21,7 +21,7 @@ void PhysicsEngine::unregisterComponent(engine::component::PhysicsComponent* com
     spdlog::trace("物理组件注销完成。");
 }
 
-void PhysicsEngine::update(double deltaTime)
+void PhysicsEngine::update(float deltaTime)
 {
     // 遍历所有注册的物理组件
     for (auto* physicsComponent : m_components) {
